Add LentOutfitCard and refactor MyOutfits/Requests
- Introduce LentOutfitCard component to contain outfit card UI and actions and refactor MyOutfits to use LentOutfitCard
- Enhance Requests page to show a simple stats card and page layout.
